Update "friends_quotes.csv" path.
I updated the summary and histogram modules. Swapped out the bar chart for a scatter plot and swapped out the box plot for a violin plot. The network graph doesn't seem to be working so we can remove it if you like. I updated all the names in the ui and server part of the code and everything seems to be working well.